rule vars: move conf var check and tests
